# Gzipping the _rates